Replace WikiTransformation with original.
